Ending of the last class

+++ Matrix.h
@@ -6,7 +6,7 @@
 
 const int MAX_COL = 4;
 
-double sumOfRow(const float mat[][MAX_COL], const int row, const int maxRow);
-double sumOfCol(const float mat[][MAX_COL], const int column, const int maxRow);
+float sumOfRow(const float mat[][MAX_COL], const int row, const int maxRow);
+float sumOfCol(const float mat[][MAX_COL], const int column, const int maxRow);
 void fillWithRandomNum(float mat[][MAX_COL], const int maxRow);
 void printMatrix(const float mat[][MAX_COL], const int maxRow);
